Start introducing chainid based functionality

+++ scripts/stop-blockchain-client.sh
@@ -1,6 +1,7 @@
 #!/usr/bin/env bash
 
-pid=$(lsof -i:8545 -sTCP:LISTEN -t); 
+pid=$(lsof -i:8545 -sTCP:LISTEN -t);
 
-echo "Killing blockchain client process $pid on port 8545"
+if [ -z "$pid" ]; then exit; else echo "Killing blockchain client process $pid on port 8545"; fi
+
 kill -INT $pid
